Revert "LPS-56219 Move"
This reverts commit ed3204fae38bbe9b5f99029ebbf14f162ccea2f1.
